// src/components/CredibilityResult.module.css.d.ts
declare const styles: {
  credibilityResult: string;
  "level-strong": string;
  "level-moderate": string;
  "level-weak": string;
  header: string;
  title: string;
  factionScores: string;
  row: string;
  factionName: string;
  score: string;
  footer: string;
  finalRow: string;
  finalLabel: string;
  finalScore: string;
};

export = styles;